wip ledger

+++ libraries/ness/sealing/test/io/intino/alexandria/sealing/LedSessionManager_.java
@@ -5,13 +5,13 @@ import io.intino.alexandria.Timetag;
 import io.intino.alexandria.datalake.file.FileDatalake;
 import io.intino.alexandria.ingestion.LedSession;
 import io.intino.alexandria.ingestion.SessionHandler;
-import io.intino.alexandria.led.LedReader;
-import io.intino.alexandria.led.PrimaryLed;
+import io.intino.alexandria.led.IndexedLed;
+import io.intino.alexandria.led.Led;
+import io.intino.alexandria.led.io.LedReader;
 import org.junit.After;
 import org.junit.Test;
 
 import java.io.File;
-import java.io.IOException;
 import java.time.LocalDateTime;
 import java.util.List;
 
@@ -24,20 +24,19 @@ public class LedSessionManager_ {
 	private static final File DATALAKE = new File("temp/datalake");
 
 	@Test
-	public void should_create_and_seal_led_session() throws IOException {
+	public void should_create_and_seal_led_session() {
 		SessionHandler handler = new SessionHandler(LOCAL_STAGE);
 		LocalDateTime dateTime = LocalDateTime.of(2019, 2, 28, 16, 15);
 		Timetag timetag = new Timetag(dateTime, Scale.Day);
 		LedSession session = handler.createLedSession();
 		List<TestSchema> stored = TestSchema.unsortedList();
-		PrimaryLed<TestSchema> testLed = PrimaryLed.load(stored.stream());
+		Led<TestSchema> testLed = IndexedLed.of(stored);
 		session.put("tank1", timetag, testLed);
 		handler.pushTo(STAGE_FOLDER);
 		new FileSessionSealer(new FileDatalake(DATALAKE), STAGE_FOLDER).seal();
-		PrimaryLed<TestSchema> led = new LedReader(new File("temp/datalake/ledger/tank1/" + timetag.value() + ".led")).read(TestSchema.class);
-		assertThat(led.size()).isEqualTo(stored.size());
+		IndexedLed<TestSchema> led = new LedReader(new File("temp/datalake/ledger/tank1/" + timetag.value() + ".led")).read(TestSchema::new);
 		assertThat(led.iterator().next().id()).isEqualTo(testLed.iterator().next().id());
-		assertThat(led.iterator().next().i()).isEqualTo(testLed.iterator().next().i());
+		assertThat(led.iterator().next().e()).isEqualTo(testLed.iterator().next().e());
 		assertThat(led.iterator().next().d()).isEqualTo(testLed.iterator().next().d());
 	}
 
@@ -48,11 +47,7 @@ public class LedSessionManager_ {
 
 	private void deleteDirectory(File directoryToBeDeleted) {
 		File[] allContents = directoryToBeDeleted.listFiles();
-		if (allContents != null) {
-			for (File file : allContents) {
-				deleteDirectory(file);
-			}
-		}
+		if (allContents != null) for (File file : allContents) deleteDirectory(file);
 		directoryToBeDeleted.delete();
 	}
 }
